send session checksum with point source update
