+++ client/src/SessionView/Overview/PointsSourceHelper/Toggle.js
