+++ client/src/SessionView/Overview/PointsSourceHelper/index.js
